pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-server = "setuptools.build_meta"

[project]
name = "server"
version = "0.3.2"
description = "A 'hello world' style Flask web server application that applies good practices."
readme = "README.md"
authors = [
    { name = "Stephen Holsapple", email = "sholsapp@gmail.com" }
]
requires-python = ">=3.11"
urls = { "Homepage" = "http://www.flask.com" }
license = "MIT"

dependencies = [
    "Authlib",
    "APScheduler",
    "Flask",
    "Flask-Bootstrap",
    "Flask-Cors",
    "Flask-REST-JSONAPI",
    "Flask-SQLALchemy",
    "Flask-Security",
    "Flask-Admin[sqlalchemy]",
    "backoff",
    "configobj",
    "gunicorn",
    "requests",
    "gevent>=25.9.1",
    "psycopg2-binary",
    "flask-mail>=0.10.0",
    "flasgger>=0.9.7.1",
    "python-dotenv>=1.2.2",
    "flask-socketio>=5.6.1",
    "gevent-websocket>=0.10.1",
]

[project.scripts]
server-worker = "serverlearnify.worker:main"

[dependency-groups]
dev = [
    "ruff",
    "mypy",
    "pytest",
    "pytest-flask",
    "alembic>=1.18.4",
    "black>=26.1.0",
]

[tool.mypy]
ignore_missing_imports = true

[tool.setuptools]
packages = ["server"]

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = [
    "F", # flake8 error
    "E", # pycodestyle error
    "W", # pycodestyle warning
    "I", # isort imports
]
